% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dist.R
\name{dfm.tokens_with_dist}
\alias{dfm.tokens_with_dist}
\title{Create a document-feature matrix}
\usage{
\method{dfm}{tokens_with_dist}(
  x,
  remove_docvars_dist = TRUE,
  weight_function = function(x) {
     1/x
 },
  ...
)
}
\arguments{
\item{x}{output of \code{\link[=tokens_dist]{tokens_dist()}}}

\item{remove_docvars_dist}{boolean, remove the "dist" document variable}

\item{weight_function}{a weight function, default to invert distance}
}
\description{
Construct a sparse document-feature matrix from the output of \code{\link[=tokens_dist]{tokens_dist()}}.
}
\details{
By default, words closer to targets are weighted higher. You might change that with another \code{weight_function}.
}
\examples{
library(quanteda)
ukimg_eu <- data_char_ukimmig2010 \%>\% tokens(remove_punct = TRUE) \%>\%
tokens_tolower() \%>\% tokens_dist(c("eu", "europe", "european"))
ukimg_eu \%>\% dfm() \%>\% dfm_select(c("immig*", "migr*")) \%>\% rowSums() \%>\% sort()
## Words further away from targets are weighted higher
ukimg_eu \%>\% dfm(weight_function = identity) \%>\% dfm_select(c("immig*", "migr*")) \%>\% rowSums() \%>\% sort()
ukimg_eu \%>\% dfm(weight_function = function(x) {1 / x^2}) \%>\%
dfm_select(c("immig*", "migr*")) \%>\% rowSums() \%>\% sort()
}
