% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dist.R
\name{dfm.tokens_with_proximity}
\alias{dfm.tokens_with_proximity}
\title{Create a document-feature matrix}
\usage{
\method{dfm}{tokens_with_proximity}(
  x,
  tolower = TRUE,
  remove_padding = FALSE,
  verbose = quanteda::quanteda_options("verbose"),
  remove_docvars_proximity = TRUE,
  weight_function = function(x) {
     1/x
 },
  ...
)
}
\arguments{
\item{x}{output of \code{\link[=tokens_proximity]{tokens_proximity()}}.}

\item{tolower}{convert all features to lowercase.}

\item{remove_padding}{logical; if \code{TRUE}, remove the "pads" left as empty tokens after calling \code{\link[quanteda:tokens]{quanteda::tokens()}} or \code{\link[quanteda:tokens_select]{quanteda::tokens_remove()}} with \code{padding = TRUE}.}

\item{verbose}{display messages if \code{TRUE}.}

\item{remove_docvars_proximity}{logical, remove the "proximity" document variable.}

\item{weight_function}{a weight function, default to invert distance,}

\item{...}{not used.}
}
\value{
a \link[quanteda:dfm-class]{quanteda::dfm} object
}
\description{
Construct a sparse document-feature matrix from the output of \code{\link[=tokens_proximity]{tokens_proximity()}}.
}
\details{
By default, words closer to keywords are weighted higher. You might change that with another \code{weight_function}.
}
\examples{
library(quanteda)
tok1 <- data_char_ukimmig2010 \%>\%
    tokens(remove_punct = TRUE) \%>\%
    tokens_tolower() \%>\%
    tokens_proximity(c("eu", "europe", "european"))
tok1 \%>\%
    dfm() \%>\%
    dfm_select(c("immig*", "migr*")) \%>\%
    rowSums() \%>\%
    sort()
## Words further away from keywords are weighted higher
tok1 \%>\%
    dfm(weight_function = identity) \%>\%
    dfm_select(c("immig*", "migr*")) \%>\%
    rowSums() \%>\%
    sort()
tok1 \%>\%
    dfm(weight_function = function(x) {
        1 / x^2
    }) \%>\%
    dfm_select(c("immig*", "migr*")) \%>\%
    rowSums() \%>\%
    sort()
}
